requirements.txt
Django==5.0.*
djangorestframework==3.15.*
django-environ==0.11.*
psycopg[binary,pool]==3.1.*
numpy==1.26.*
django-cors-headers==4.3.*
python-dateutil==2.9.*
drf-nested-routers==0.93.*
httpx==0.27.*
# gunicorn==21.2.*
django-filter==24.*
drf-spectacular==0.27.*
djangorestframework-simplejwt==5.3.*
pymodbus==3.6.*
pandas==2.2.*

# Developer Tools
# ------------------------------------------------------------------------------
django_extensions==3.2.*
Werkzeug
django-debug-toolbar==4.1.*
rich==13.7.*
ipdb==0.13.*
ipython==8.23.*
mkdocs==1.5.*
flake8==7.0.*
ruff==0.4.*


# Testing
# ------------------------------------------------------------------------------
coverage==7.2.*
pytest==8.2.*
pytest-cov==5.0.*
